requirements.txt
langchain~=0.3.7
langchain-community~=0.3.5
langchain-openai
qdrant-client~=1.12.1
pypdf
langchain-qdrant
sentence-transformers~=3.2.1
anthropic~=0.39.0
google-generativeai
python-dotenv~=1.0.1
requests~=2.32.3
protobuf~=5.28.3
pytesseract~=0.3.13
pillow~=10.4.0
pandas~=2.2.3